requirements.txt
https://www.django-rest-framework.org/tutorial/quickstart/#project-setup
pip install django
pip install djangorestframework
django-admin startproject tutorial .
cd tutorial
django-admin startapp quickstart
python manage.py migrate
python manage.py createsuperuser
u: msb
p: 123
python manage.py runserver

1. Create File quickstart/serializers.py

from django.contrib.auth.models import User, Group
from rest_framework import serializers
from .models import Movie

class MovieSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Movie
        fields = ('id', 'title', 'desc', 'year')

2. Copy to Views of quickstart

from django.contrib.auth.models import User, Group
from rest_framework import viewsets
from .serializers import MovieSerializer
from .models import Movie

class MovieViewSet(viewsets.ModelViewSet):
    queryset = Movie.objects.all()
    serializer_class = MovieSerializer

3. Copy to urls of tutorial

from django.contrib import admin
from django.urls import include, path
from rest_framework import routers
from tutorial.quickstart import views

router = routers.DefaultRouter()
router.register(r'movies', views.MovieViewSet)

urlpatterns = [
    path('admin/', admin.site.urls),
    path('', include(router.urls)),
    path('api-auth/', include('rest_framework.urls', namespace='rest_framework'))
]

4. Add in Installed Apps :     'tutorial.quickstart',

5. Create Models / DB

from django.db import models

# Create your models here.
class Movie(models.Model):
    title = models.CharField(max_length=40)
    desc = models.CharField(max_length=100)
    year = models.IntegerField()

6. Add in admin.py

from .models import Movie
admin.site.register(Movie)


# Now For Angular

1. create frontend folder similar to tutorial
2. ng new curd
3. cd curd
4. ng serve
5. Add to app.component.html
<h2>Movies List</h2>
<ul>
    <li *ngFor="let movie of movies">
        <h2>{{ movie.title }}</h2>
    </li>
</ul>

6. Add into app.component.ts
export class AppComponent {
  title = 'curd';
  movies = [{title: 'titanic'}, {title:'Champa'}];
}

7. ng generate service quickstart
8. Open quickstart.service.ts

import { Injectable } from '@angular/core';
import { HttpClient, HttpHandler, HttpHeaders } from '@angular/common/http';
import { observable, Observable } from 'rxjs'

@Injectable({
  providedIn: 'root'
})
export class QuickstartService {

  baseurl = "http://127.0.0.1:8000/"
  httpHeaders = new HttpHeaders({'Content-Type':'application/json'})

  constructor(private http: HttpClient) { }

  getAllMovies():Observable<any>{
    return this.http.get(this.baseurl + 'movies/', {headers:this.httpHeaders});
  }
}


9. open app.component.ts
import { Component } from '@angular/core';
import { QuickstartService } from './quickstart.service';

@Component({
  selector: 'app-root',
  templateUrl: './app.component.html',
  styleUrls: ['./app.component.css'],
  providers: [QuickstartService]
})
export class AppComponent {
  title = 'curd';

  movies = [{title: 'titanic'}, {title:'Champa'}];

  constructor(private quickstart:QuickstartService){
    this.getMovies();
  }
  getMovies = () => {
    this.quickstart.getAllMovies().subscribe(
      data => {
        this.movies = data;
      },
      error => {
        console.log(error);
      },

    )
  }
}

10. import in app.module.ts

import { BrowserModule } from '@angular/platform-browser';
import { NgModule } from '@angular/core';
import { HttpClientModule } from '@angular/common/http'
import { AppComponent } from './app.component';
import { from } from 'rxjs';

@NgModule({
  declarations: [
    AppComponent
  ],
  imports: [
    BrowserModule,
    HttpClientModule
  ],
  providers: [],
  bootstrap: [AppComponent]
})
export class AppModule { }

11. pip install django-cors-headers (https://pypi.org/project/django-cors-headers/)

12. then add it to your installed apps: (django)

INSTALLED_APPS = [
    ...
    'corsheaders',
    ...
]

13. You will also need to add a middleware class to listen in on responses: (django)

MIDDLEWARE = [  # Or MIDDLEWARE_CLASSES on Django < 1.10
    ...
    'corsheaders.middleware.CorsMiddleware',
    'django.middleware.common.CommonMiddleware',
    ...
]

14. Add it after Allowed Hosts
CORS_ORIGIN_WHITELIST = [
    'http://localhost:4200'
]

15. Done
